adding some stuff

+++ src/main/java/com/codingtech/formationcenter/repo/DeveloperRepo.java
@@ -3,5 +3,5 @@ package com.codingtech.formationcenter.repo;
 import com.codingtech.formationcenter.entity.Developer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeveloperRepo  extends JpaRepository<Developer,Integer> {
+public interface DeveloperRepo  extends JpaRepository<Developer,Long> {
 }
