+++ src/main/java/com/codingtech/formationcenter/repo/EducationRepo.java
@@ -3,5 +3,5 @@ package com.codingtech.formationcenter.repo;
 import com.codingtech.formationcenter.entity.Education;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EducationRepo  extends JpaRepository<Education,Integer> {
+public interface EducationRepo  extends JpaRepository<Education,Long> {
 }
