+++ src/main/java/com/codingtech/formationcenter/repo/ExperienceRepo.java
@@ -3,5 +3,5 @@ package com.codingtech.formationcenter.repo;
 import com.codingtech.formationcenter.entity.Experience;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExperienceRepo  extends JpaRepository<Experience,Integer> {
+public interface ExperienceRepo  extends JpaRepository<Experience,Long> {
 }
