+++ src/main/java/com/codingtech/formationcenter/repo/ProjectRepo.java
@@ -3,5 +3,5 @@ package com.codingtech.formationcenter.repo;
 import com.codingtech.formationcenter.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepo  extends JpaRepository<Project,Integer> {
+public interface ProjectRepo  extends JpaRepository<Project,Long> {
 }
